+++ src/menu.tcl
@@ -518,11 +518,12 @@ proc menu::FillMenu {} {
   ### ___________________ Check, File list __________________ ###
 
   $m add separator
-  $m add command -label $al(MC,checktcl...) -command alited::CheckRun
   menu $m.filelist -tearoff 0
   $m add cascade -label $al(MC,filelist) -menu $m.filelist
   $m.filelist add command -label $al(MC,filelist) -command {alited::bar::BAR popList} -accelerator $al(acc_21)
   $m.filelist add checkbutton -label [msgcat::mc {Sorted}] -variable ::alited::al(sortList)
+  $m add command -label $al(MC,checktcl...) -command alited::CheckRun
+  $m add command -label [msgcat::mc {Project Printer...}] -command alited::PrinterRun
 
   ### ________________________ Paver _________________________ ###